refactor(api: profile routes): add check before return response and fix catch error for all http requests

+++ app/api/profile/[id]/route.js
@@ -1,39 +1,66 @@
 import Profile from "@modules/server/models/profile";
 import connectDb from "lib/mongodb";
 import { NextResponse } from "next/server";
+import checkId from "@modules/server/utils/checkId";
 
 export async function PATCH(request, { params }) {
   try {
     const { id } = params;
+    if (!checkId(id)) {
+      throw { message: "Wrong id", status: 500 };
+    }
     const profile = await request.json();
+    if (!profile) {
+      throw { message: "Not found", status: 400 };
+    }
     await connectDb();
-    const updatedProfile = await Profile.findByIdAndUpdate(id, profile, { runValidators: true });
-    return NextResponse.json(updatedProfile, { status: 200 });
+    const updatedProfile = await Profile.findByIdAndUpdate(id, profile, {
+      runValidators: true,
+    });
+    return NextResponse.json(
+      updatedProfile,
+      { message: "Profile updated" },
+      { status: 202 }
+    );
   } catch (err) {
-    const { errors } = err;
-    return NextResponse.json(errors, { status: 404 });
+    const { message, status } = err;
+    return NextResponse.json({ message, status }, { status: status || 404 });
   }
 }
 
 export async function DELETE(request, { params }) {
   try {
-    const { id } = params
+    const { id } = params;
+    if (!checkId(id)) {
+      throw { message: "Wrong id", status: 500 };
+    }
     await connectDb();
     const deletedProfile = await Profile.findByIdAndDelete(id);
-    return NextResponse.json(deletedProfile)
+    return NextResponse.json(
+      deletedProfile,
+      { message: "Profile deleted" },
+      { status: 202 }
+    );
   } catch (err) {
-    return NextResponse.json({message: "Error"}, {status: 404});
+    const { message, status } = err;
+    return NextResponse.json({ message, status }, { status: status || 404 });
   }
 }
 
 export async function GET(request, { params }) {
   try {
     const { id } = params;
+    if (!checkId(id)) {
+      throw { message: "Wrong id", status: 500 };
+    }
     await connectDb();
     const profile = await Profile.findById(id);
-    return NextResponse.json(profile, {status: 200});
-  } catch(err) {
-    const { errors } = err;
-    return NextResponse.json(errors, {status: 404});
+    if (!profile) {
+      throw { message: "Not found", status: 400 };
+    }
+    return NextResponse.json(profile, { status: 200 });
+  } catch (err) {
+    const { message, status } = err;
+    return NextResponse.json({ message, status }, { status: status || 404 });
   }
 }
